Moved routes for quizinfo, skip, previous and all other routes for answering a quiz to answering_quiz blueprint
