+++ api/templates/finishquiz.html
@@ -31,7 +31,7 @@
     
     <div class="results-actions">
       <a href="{{ url_for('dash.home') }}" class="button return-home">Return Home</a>
-      <a href="{{ url_for('quiz.quizinfo', quiz_id=quiz_id) }}" class="button retry-quiz">Retry</a>
+      <a href="{{ url_for('taking.quizinfo', quiz_id=quiz_id) }}" class="button retry-quiz">Retry</a>
     </div>
   </div>
 </main>
